requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.8
openai==2.15.0
